requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --generate-hashes
#
attrs==19.3.0 \
    --hash=sha256:08a96c641c3a74e44eb59afb61a24f2cb9f4d7188748e76ba4bb5edfa3cb7d1c \
    --hash=sha256:f7b7ce16570fe9965acd6d30101a28f62fb4a7f9e926b3bbc9b61f8b04247e72 \
    # via jsonschema, pytest
cython==0.29.14 \
    --hash=sha256:03f6bbb380ad0acb744fb06e42996ea217e9d00016ca0ff6f2e7d60f580d0360 \
    --hash=sha256:05e8cfd3a3a6087aec49a1ae08a89171db991956209406d1e5576f9db70ece52 \
    --hash=sha256:05eb79efc8029d487251c8a2702a909a8ba33c332e06d2f3980866541bd81253 \
    --hash=sha256:094d28a34c3fa992ae02aea1edbe6ff89b3cc5870b6ee38b5baeb805dc57b013 \
    --hash=sha256:0c70e842e52e2f50cc43bad43b5e5bc515f30821a374e544abb0e0746f2350ff \
    --hash=sha256:1dcdaa319558eb924294a554dcf6c12383ec947acc7e779e8d3622409a7f7d28 \
    --hash=sha256:1fc5bdda28f25fec44e4721677458aa509d743cd350862270309d61aa148d6ff \
    --hash=sha256:280573a01d9348d44a42d6a9c651d9f7eb1fe9217df72555b2a118f902996a10 \
    --hash=sha256:298ceca7b0f0da4205fcb0b7c9ac9e120e2dafffd5019ba1618e84ef89434b5a \
    --hash=sha256:4074a8bff0040035673cc6dd365a762476d6bff4d03d8ce6904e3e53f9a25dc8 \
    --hash=sha256:41e7068e95fbf9ec94b41437f989caf9674135e770a39cdb9c00de459bafd1bc \
    --hash=sha256:47e5e1502d52ef03387cf9d3b3241007961a84a466e58a3b74028e1dd4957f8c \
    --hash=sha256:521340844cf388d109ceb61397f3fd5250ccb622a1a8e93559e8de76c80940a9 \
    --hash=sha256:6c53338c1811f8c6d7f8cb7abd874810b15045e719e8207f957035c9177b4213 \
    --hash=sha256:75c2dda47dcc3c77449712b1417bb6b89ec3b7b02e18c64262494dceffdf455e \
    --hash=sha256:773c5a98e463b52f7e8197254b39b703a5ea1972aef3a94b3b921515d77dd041 \
    --hash=sha256:78c3068dcba300d473fef57cdf523e34b37de522f5a494ef9ee1ac9b4b8bbe3f \
    --hash=sha256:7bc18fc5a170f2c1cef5387a3d997c28942918bbee0f700e73fd2178ee8d474d \
    --hash=sha256:7f89eff20e4a7a64b55210dac17aea711ed8a3f2e78f2ff784c0e984302583dd \
    --hash=sha256:89458b49976b1dee5d89ab4ac943da3717b4292bf624367e862e4ee172fcce99 \
    --hash=sha256:986f871c0fa649b293061236b93782d25c293a8dd8117c7ba05f8a61bdc261ae \
    --hash=sha256:a0f495a4fe5278aab278feee35e6102efecde5176a8a74dd28c28e3fc5c8d7c7 \
    --hash=sha256:a14aa436586c41633339415de82a41164691d02d3e661038da533be5d40794a5 \
    --hash=sha256:b8ab3ab38afc47d8f4fe629b836243544351cef681b6bdb1dc869028d6fdcbfb \
    --hash=sha256:bb487881608ebd293592553c618f0c83316f4f13a64cb18605b1d2fb9fd3da3e \
    --hash=sha256:c0b24bfe3431b3cb7ced323bca813dbd13aca973a1475b512d3331fd0de8ec60 \
    --hash=sha256:c7894c06205166d360ab2915ae306d1f7403e9ce3d3aaeff4095eaf98e42ce66 \
    --hash=sha256:d4039bb7f234ad32267c55e72fd49fb56078ea102f9d9d8559f6ec34d4887630 \
    --hash=sha256:e4d6bb8703d0319eb04b7319b12ea41580df44fd84d83ccda13ea463c6801414 \
    --hash=sha256:e8fab9911fd2fa8e5af407057cb8bdf87762f983cba483fa3234be20a9a0af77 \
    --hash=sha256:f3818e578e687cdb21dc4aa4a3bc6278c656c9c393e9eda14dd04943f478863d \
    --hash=sha256:fe666645493d72712c46e4fbe8bec094b06aec3c337400479e9704439c9d9586 \
    # via pyjnius
importlib-metadata==1.5.0 \
    --hash=sha256:06f5b3a99029c7134207dd882428a66992a9de2bef7c2b699b5641f9886c3302 \
    --hash=sha256:b97607a1a18a5100839aec1dc26a1ea17ee0d93b20b0f008d80a5a050afb200b \
    # via jsonschema, pluggy, pytest
jsonschema==3.2.0 \
    --hash=sha256:4e5b3cf8216f577bee9ce139cbe72eca3ea4f292ec60928ff24758ce626cd163 \
    --hash=sha256:c8a85b28d377cc7737e46e2d9f2b4f44ee3c0e1deac6bf46ddefc7187d30797a
more-itertools==8.2.0 \
    --hash=sha256:5dd8bcf33e5f9513ffa06d5ad33d78f31e1931ac9a18f33d37e77a180d393a7c \
    --hash=sha256:b1ddb932186d8a6ac451e1d95844b382f55e12686d51ca0c68b6f61f2ab7a507 \
    # via pytest
packaging==20.1 \
    --hash=sha256:170748228214b70b672c581a3dd610ee51f733018650740e98c7df862a583f73 \
    --hash=sha256:e665345f9eef0c621aa0bf2f8d78cf6d21904eef16a93f020240b704a57f1334 \
    # via pytest
pluggy==0.13.1 \
    --hash=sha256:15b2acde666561e1298d71b523007ed7364de07029219b604cf808bfa1c765b0 \
    --hash=sha256:966c145cd83c96502c3c3868f50408687b38434af77734af1e9ca461a4081d2d \
    # via pytest
py==1.8.1 \
    --hash=sha256:5e27081401262157467ad6e7f851b7aa402c5852dbcb3dae06768434de5752aa \
    --hash=sha256:c20fdd83a5dbc0af9efd622bee9a5564e278f6380fffcacc43ba6f43db2813b0 \
    # via pytest
pyjnius==1.2.1 \
    --hash=sha256:100057323cec78eda64b7fe629d8bf0fc09687cacef24cd25e983367d467ba7a \
    --hash=sha256:16305191a05b71ef565baf8cd8945f88d85f8f4b8aacecdfbcdd5f4f09853280 \
    --hash=sha256:19bc5135b7d9e791d8950dde8d9a258d22653080f9342d194aa7cf6f7181217e \
    --hash=sha256:1aada711a06b0c11c33a215e4d5abad94e45e7a0ec4c02cfc2aa31c018bdc95e \
    --hash=sha256:1f6701596218421c02a18050983009b2e54e7b0e3c564d5e98a486e18cfeb47e \
    --hash=sha256:37897af8b5fbff19ce10ca2870d29152aee8f5e1f77b4bcd561b9208c36c4e4e \
    --hash=sha256:3a49fde713b9504092886097962f8c6a3dfe3ab87ee59f1ddd7a02baa6263e24 \
    --hash=sha256:3c770933f3f9cb8abf3871a8fd91ae23e93276deef6f6ef518bbdb62fb4af3c5 \
    --hash=sha256:3ee9a7d3a0a7cba021e44d58f8ba49adb165499b326de137abd14099a5ad87e1 \
    --hash=sha256:4d4b5ee12c0ab420131f844e838e9e171efaa9834c86a183391d87fb08f24395 \
    --hash=sha256:4e6731fba2f98cb7e87666edd38559b88261a7dcb66c24bfa7ace8b23a182914 \
    --hash=sha256:5d7cda4c1ccc6b7c1eca8d60e0bd06bc825588aecc3e882943f4a79d854fca6d \
    --hash=sha256:60df1b81af2d453152792dd153914500b062f15f44159c6a39ee10b2576cae6c \
    --hash=sha256:6f82c06ab2946d190987847aaa21f3dc2df49cd776da24c924544c4a01797575 \
    --hash=sha256:7022591c1e38affe881a2311d377694edb42589c582c52d377c2c4041d85dd23 \
    --hash=sha256:83d3300a7352b9e6fd555d2fecd4831f1d28d9f1c055ab7af11ebc68b98050ab \
    --hash=sha256:87f54b62581d32b512add08127a11f8ace0a6f042d7d650127e85a7ee51ac8f7 \
    --hash=sha256:8bc4e6499b29b74e3b38614fdc1643afa17dfa4755e24682c8f829106b4d51a2 \
    --hash=sha256:9ecfa2286234b54f09d9141afc649ba3e77e4efe0303fb8a40a90397fa716c0f \
    --hash=sha256:ced833dbac7660a0ec9ca505bafa2fa094c68d15f25a5c2848b614ccd6fa260a \
    --hash=sha256:de0e7251c306faf960d37d31a588edd3e10cfa6edcae91822d72a9f47b9739b3 \
    --hash=sha256:ef8acab01f9a23f35a3f202ca4fdb3f1f2f3f81754b046ae1e8fc1850d53dc0b
pyparsing==2.4.6 \
    --hash=sha256:4c830582a84fb022400b85429791bc551f1f4871c33f23e44f353119e92f969f \
    --hash=sha256:c342dccb5250c08d45fd6f8b4a559613ca603b57498511740e65cd11a2e7dcec \
    # via packaging
pyrsistent==0.15.7 \
    --hash=sha256:cdc7b5e3ed77bed61270a47d35434a30617b9becdf2478af76ad2c6ade307280 \
    # via jsonschema
pytest==5.3.5 \
    --hash=sha256:0d5fe9189a148acc3c3eb2ac8e1ac0742cb7618c084f3d228baaec0c254b318d \
    --hash=sha256:ff615c761e25eb25df19edddc0b970302d2a9091fbce0e7213298d85fb61fef6
six==1.14.0 \
    --hash=sha256:236bdbdce46e6e6a3d61a337c0f8b763ca1e8717c03b369e87a7ec7ce1319c0a \
    --hash=sha256:8f3cd2e254d8f793e7f3d6d9df77b92252b52637291d0f0da013c76ea2724b6c \
    # via jsonschema, packaging, pyjnius, pyrsistent
wcwidth==0.1.8 \
    --hash=sha256:8fd29383f539be45b20bd4df0dc29c20ba48654a41e661925e612311e9f3c603 \
    --hash=sha256:f28b3e8a6483e5d49e7f8949ac1a78314e740333ae305b4ba5defd3e74fb37a8 \
    # via pytest
zipp==2.2.0 \
    --hash=sha256:5c56e330306215cd3553342cfafc73dda2c60792384117893f3a83f8a1209f50 \
    --hash=sha256:d65287feb793213ffe11c0f31b81602be31448f38aeb8ffc2eb286c4f6f6657e \
    # via importlib-metadata

# WARNING: The following packages were not pinned, but pip requires them to be
# pinned when the requirements file includes hashes. Consider using the --allow-unsafe flag.
# setuptools
